+++ src/index.html
@@ -57,6 +57,9 @@
         <p class="message message-play hidden">Could you reach the 2048?</p>
         <p class="message message-win hidden">Winner! Congrats! You did it!</p>
         <p class="message message-start">Press "Start" to begin game. Good luck!</p>
+        <p class="message message-score">
+          Highest score: <span class="highest-score">0</span>
+        </p>
       </div>
     </div>
     <script type="text/javascript" src="scripts/main.js"></script>
