+++ src/styles/main.scss
@@ -183,6 +183,11 @@ h1 {
   color: #f9f6f2;
 }
 
+.highest-score {
+  color: #616161;
+  border-bottom: 1px solid #776e65;
+}
+
 .message-container {
   width: 100%;
   height: 150px;
